Merge main into dev

+++ frontend/src/components/GamemasterMenu.tsx
@@ -2,29 +2,31 @@
 
 import { useState, RefObject } from 'react';
 import { useAuthedFetch } from '@/hooks/useAuthedFetch';
+import ReconnectingWebSocket from "reconnecting-websocket";
 import { RoleInstance } from '@/lib/Types';
+
 
 interface GameMasterMenuProps {
     joinCode: string;
-    socketRef: RefObject<WebSocket | null>;
+    socketRef: RefObject<ReconnectingWebSocket | null>;
     socketReady: boolean;
-    roleInstance: RoleInstance;
-}
+    roleInstance: RoleInstance | null;
+};
 
 export default function GamemasterMenu({ joinCode, socketRef, socketReady, roleInstance }: GameMasterMenuProps) {
-
     const authedFetch = useAuthedFetch();
 
+    const [open, setOpen] = useState<boolean>(true);
     const [deletingGame, setDeletingGame] = useState<boolean>(false);
 
     const handleDeleteGame = async () => {
-        if (!socketReady || !socketRef.current) return;
+        if (!joinCode || !socketReady || !socketRef.current || !roleInstance) return;
         if (!confirm("Are you sure you want to delete this game?\nThis is irreversible.")) return;
         const socket = socketRef.current;
 
         try {
             setDeletingGame(true);
-            const res = await authedFetch(`/api/game-instances/delete/${roleInstance.team_instance.game_instance.join_code}/`, {
+            const res = await authedFetch(`/api/game-instances/delete/${roleInstance?.team_instance.game_instance.join_code}/`, {
                 method: 'DELETE'
             });
 
@@ -42,34 +44,46 @@ export default function GamemasterMenu({ joinCode, socketRef, socketReady, roleI
                     }
                 }));
             }
-            
+
         } catch (err: unknown) {
             console.error(err);
+
             if (err instanceof Error) {
                 alert(err.message);
             }
+
             setDeletingGame(false);
         }
     };
 
     return (
         <div className="bg-neutral-700 rounded-lg p-4 mt-4 text-white">
-            <h3 className="text-lg font-semibold mb-2">Gamemaster Menu</h3>
-            {/* Delete Game */}
-            <div>
+            <div className="flex justify-between items-center mb-2">
+                <h3 className="text-lg font-semibold">Gamemaster Menu</h3>
                 <button
-                    onClick={handleDeleteGame}
-                    disabled={deletingGame}
-                    className={`px-3 py-2 rounded transition
-                        ${deletingGame
-                            ? "bg-gray-500 cursor-not-allowed"
-                            : "bg-red-600 cursor-pointer hover:bg-red-500"
-                        }
-                    `}
+                    onClick={() => setOpen(!open)}
+                    className="text-sm bg-neutral-600 px-2 py-1 rounded cursor-pointer hover:bg-neutral-500"
                 >
-                    {deletingGame ? "Deleting Game..." : "Delete Game"}
+                    {open ? '-' : '+'}
                 </button>
             </div>
+            {open && (
+                // Delete Game
+                <div>
+                    <button
+                        onClick={() => handleDeleteGame()}
+                        disabled={deletingGame}
+                        className={`px-3 py-2 rounded
+                            ${deletingGame
+                                ? "bg-gray-500 cursor-not-allowed"
+                                : "bg-red-600 cursor-pointer hover:bg-red-500"
+                            }
+                        `}
+                    >
+                        {deletingGame ? "Deleting Game..." : "Delete Game"}
+                    </button>
+                </div>
+            )}
         </div>
     );
 };
